Refactor: refactored code
