+++ pkg/network/protocols/core/protocol.go
@@ -102,26 +102,6 @@ func (p *Protocol) RequestAttestations(id iotago.CommitmentID, to ...network.Pee
 	}}}, protocolID, to...)
 }
 
-func (p *Protocol) SendWarpSyncResponse(id iotago.CommitmentID, blockIDs iotago.BlockIDs, merkleProof *merklehasher.Proof[iotago.Identifier], to ...network.PeerID) {
-	serializer := p.apiProvider.APIForSlot(id.Index())
-
-	p.network.Send(&nwmodels.Packet{Body: &nwmodels.Packet_WarpSyncResponse{
-		WarpSyncResponse: &nwmodels.WarpSyncResponse{
-			CommitmentId: lo.PanicOnErr(id.Bytes()),
-			BlockIds:     lo.PanicOnErr(serializer.Encode(blockIDs)),
-			MerkleProof:  lo.PanicOnErr(json.Marshal(merkleProof)),
-		},
-	}}, protocolID, to...)
-}
-
-func (p *Protocol) SendWarpSyncRequest(id iotago.CommitmentID, to ...network.PeerID) {
-	p.network.Send(&nwmodels.Packet{Body: &nwmodels.Packet_WarpSyncRequest{
-		WarpSyncRequest: &nwmodels.WarpSyncRequest{
-			CommitmentId: lo.PanicOnErr(id.Bytes()),
-		},
-	}}, protocolID, to...)
-}
-
 func (p *Protocol) Shutdown() {
 	p.network.UnregisterProtocol(protocolID)
 
@@ -148,11 +128,9 @@ func (p *Protocol) handlePacket(nbr network.PeerID, packet proto.Message) (err e
 			p.onAttestationsRequest(packetBody.AttestationsRequest.GetCommitmentId(), nbr)
 		})
 	case *nwmodels.Packet_WarpSyncRequest:
-		p.workerPool.Submit(func() { p.onWarpSyncRequest(packetBody.WarpSyncRequest.GetCommitmentId(), nbr) })
+		p.handleWarpSyncRequest(packetBody.WarpSyncRequest.GetCommitmentId(), nbr)
 	case *nwmodels.Packet_WarpSyncResponse:
-		p.workerPool.Submit(func() {
-			p.onWarpSyncResponse(packetBody.WarpSyncResponse.GetCommitmentId(), packetBody.WarpSyncResponse.GetBlockIds(), packetBody.WarpSyncResponse.GetMerkleProof(), nbr)
-		})
+		p.handleWarpSyncResponse(packetBody.WarpSyncResponse.GetCommitmentId(), packetBody.WarpSyncResponse.GetBlockIds(), packetBody.WarpSyncResponse.GetMerkleProof(), nbr)
 	default:
 		return ierrors.Errorf("unsupported packet; packet=%+v, packetBody=%T-%+v", packet, packetBody, packetBody)
 	}
@@ -251,42 +229,6 @@ func (p *Protocol) onAttestationsRequest(commitmentIDBytes []byte, id network.Pe
 	}
 
 	p.Events.AttestationsRequestReceived.Trigger(iotago.CommitmentID(commitmentIDBytes), id)
-}
-
-func (p *Protocol) onWarpSyncRequest(commitmentIDBytes []byte, id network.PeerID) {
-	commitmentID, _, err := iotago.SlotIdentifierFromBytes(commitmentIDBytes)
-	if err != nil {
-		p.Events.Error.Trigger(ierrors.Wrap(err, "failed to deserialize commitmentID in warp sync request"), id)
-
-		return
-	}
-
-	p.Events.WarpSyncRequestReceived.Trigger(commitmentID, id)
-}
-
-func (p *Protocol) onWarpSyncResponse(commitmentIDBytes []byte, blockIDsBytes []byte, merkleProofBytes []byte, id network.PeerID) {
-	commitmentID, _, err := iotago.SlotIdentifierFromBytes(commitmentIDBytes)
-	if err != nil {
-		p.Events.Error.Trigger(ierrors.Wrap(err, "failed to deserialize commitmentID in warp sync response"), id)
-
-		return
-	}
-
-	var blockIDs iotago.BlockIDs
-	if _, err = p.apiProvider.APIForSlot(commitmentID.Index()).Decode(blockIDsBytes, &blockIDs, serix.WithValidation()); err != nil {
-		p.Events.Error.Trigger(ierrors.Wrap(err, "failed to deserialize block ids"), id)
-
-		return
-	}
-
-	merkleProof := new(merklehasher.Proof[iotago.Identifier])
-	if err = json.Unmarshal(merkleProofBytes, merkleProof); err != nil {
-		p.Events.Error.Trigger(ierrors.Wrapf(err, "failed to deserialize merkle proof when receiving attestations for commitment %s", commitmentID), id)
-
-		return
-	}
-
-	p.Events.WarpSyncResponseReceived.Trigger(commitmentID, blockIDs, merkleProof, id)
 }
 
 func newPacket() proto.Message {
